# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payroll/upload-master:
    post:
      operationId: uploadMaster
      tags: [payroll]
      summary: Upload master file
      description: Upload the master Excel file (MM РАСЧЕТ ГРАФИКА)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                month:
                  type: integer
                  minimum: 1
                  maximum: 12
              required: [file, month]
      responses:
        "200":
          description: Master file uploaded and parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MasterUploadResponse"
  /payroll/upload-reports:
    post:
      operationId: uploadReports
      tags: [payroll]
      summary: Upload payroll report files
      description: Upload one or more RESULT_PAYROLL report Excel files
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              required: [files]
      responses:
        "200":
          description: Reports uploaded and processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportsUploadResponse"
...
components:
  schemas:
...
    MasterUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        employeeCount:
          type: integer
        month:
          type: integer
        monthName:
          type: string
        normHours:
          type: integer
        hourCost:
          type: number
        employees:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeInfo"
      required: [success, message, employeeCount, month, monthName, normHours, hourCost, employees]
    EmployeeInfo:
      type: object
      properties:
        fio:
          type: string
        rawFio:
          type: string
        status:
          type: string
        dismissDate:
          type: string
          nullable: true
        existingHours:
          type: object
          additionalProperties:
            type: number
        totalExistingHours:
          type: number
      required: [fio, rawFio, status, totalExistingHours]
    ReportsUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        filesProcessed:
          type: integer
        totalRecords:
          type: integer
        reportNames:
          type: array
          items:
            type: string
        matchedEmployees:
          type: array
          items:
            type: string
        unmatchedEmployees:
          type: array
          items:
            type: string
      required: [success, message, filesProcessed, totalRecords, reportNames, matchedEmployees, unmatchedEmployees]
...