# lib/api-zod/src/generated/types/masterUploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeInfo } from "./employeeInfo";

export interface MasterUploadResponse {
  success: boolean;
  message: string;
  employeeCount: number;
  month: number;
  monthName: string;
  normHours: number;
  hourCost: number;
  employees: EmployeeInfo[];
}