# lib/api-zod/src/generated/types/reportsUploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportsUploadResponse {
  success: boolean;
  message: string;
  filesProcessed: number;
  totalRecords: number;
  reportNames: string[];
  matchedEmployees: string[];
  unmatchedEmployees: string[];
}